[Support][Telemetry] Use the EntitlementStore for the user ID. (#27825)
This patch switches the telemetry infrastructure to use the EntitlementStore to find the user ID. This patch uses an EntitlementStore passed in from the caller whenever possible, but we create TelemetryContexts all over the stack, so this patch also adds a (soon-to-be deprecated) constructor for TelemetryContext that default-constructs an EntitlementStore. This will be dangerous in the future because the user ID will no longer be in `modular.cfg`, which is what this default constructor relies on.

MODULAR_ORIG_COMMIT_REV_ID: 9d55365ced48d1049b65a32a09284520b50985a4