[Stdlib] Add a private boole method to convert bool to int in math pkg (#28117)
This adds a boole method (private for now) to convert a bool to an int.
This "in theory" does not generate a conditional because it is using a
cast operation, but in practice will lower to a select instruction by
llvm.

closes #28115

MATH_ORIG_REV_ID: c33776c96043f2994a4eb7101d250361767e3a89