[mojo-lang][LowerSemanticCF] Fix miscompilation with error handling. (#28141)
When scanning the body of a loop, we would unconditionally reset the
parent context's "can throw" bit, instead of or'ing into it.  This
caused us to incorrectly decide that the body of a try doesn't throw,
and therefore we'd delete the exception code.

This fixes #27937

MODULAR_ORIG_COMMIT_REV_ID: 6778c64efff05f86352b8f4954731658fe91ed4c