[Kernels][GPU] Switch lane_id to return a uint32, NFC (#28147)
This switches lane_id to use uint32 instead of ssize_t. The reason is
because we know that lane_id is always uint32 and we can simplify the
divide and mod operations in those cases to avoid branching and checking
for negative values.

LINALG_ORIG_REV_ID: 41040ccaf337cf4defca75bd4712f52e8511feca