[mojo-lang] Various cleanups to call emission, NFC. (#28220)
This introduces new `SharedState::typeHasMember` methods to simplify
common code that was using overload set lookup to probe for members,
and using verbose calls to `lookupAndResolveDecl` in other places.

This also converts a weird "constructor" on OverloadSet to be a
static method that returns an OverloadSet.  It is doing a somewhat
strange internal dance of overwriting itself which is unorthodox
for a ctor.

MODULAR_ORIG_COMMIT_REV_ID: 6ac73e616d4fa39739ec6ce7e4e98bdd49f12211