[mojo-lang] Refactor InputParamBindings to track its context. NFC (#28235)
This change gives a "declScope" and "SharedState" member, and plumbs
this in through all the callers. This is the only state that this
code should be using - right now it is taking an ExprEmitter which
carries information about the dynamic program as well.  This change
will allow us to remove a bunch of extra arguments from various methods
on this class and OverloadSet, and clarifies what it is allowed to
poke at.

MODULAR_ORIG_COMMIT_REV_ID: b7c902efb1f8fa7a643d3dbbb20eaa78abeea85f