[LLCL] Re-enable notion of 'current' runtime (#28280)
This re-implements the notion of the 'current' runtime for a thread,
using the `CompactRuntimePtr::{get,set}CurrentRuntime` helpers and TLS.
Unlike the impl in #27560 (which I disabled in #27928 since it was
triggering deadlocks in CI), this impl does not depend on the task
affinity API and instead just directly sets the TLS `CompactRuntimePtr`
value either when the worker thread is created, or when the 'main'
thread is about to donate itself to processing work items. This is nice
and simple.

With this the Mojo `Runtime()` ctor will correctly yield the existing
runtime, should one already be created within Mojo or by the C++
runtime, which in turn means we don't need additional plumbing when
calling Mojo `parallelize`, especially not the dreaded `out_chain`.

However, it means the `CompactRuntimePtr` for a Runtime must be
'reserved' before the `Runtime` itself is created so that it can be
passed to the work queue constructor. That's simple enough, but it
forced a cleanup of the O(100) places we currently construct a runtime
with various (and I suspect not thought out) combinations of options.
That is 100% busy work refactoring.

MODULAR_ORIG_COMMIT_REV_ID: 6ecbeb625da74a1380f8d3264c37f515baadf05c