[KGEN] Revert #28331 "More grooming for implicit lifetimes, NFC" (#28332)
This reverts the changes to not need #implicitlifetimes when there
are no ParamDecls on a signature type.  It turns out that implicit
lifetime decls are actually not stored in the signature at all, they
are stored on the LIT FuncDecl, so it is entirely valid to make a
decl with no input parameters in the signature, but that still have
implicit lifetime decls.

MODULAR_ORIG_COMMIT_REV_ID: 3b7fc174d30548a299df00c96f906659112ab0de