[mojo-lang] Move byref results to use `!lit.ref`+lifetimes. (#28337)
This is a major step forward to for the internal architecture of
the Mojo compiler, moving the by-ref result arguments to use an
implicitly declared lifetime parameter that is bound at the call
site.

This work requires getting everything that touches byref results
to cooperate, and builds on the massive lift Jeff did a ~month
ago by adding implicit lifetimes as a concept to LIT dialects.

MODULAR_ORIG_COMMIT_REV_ID: 67433c77bce7216053f6f131983a1a1579beb8c4