[mojo-lang] Improve closure conversion emission. (#28344)
This improves the type safety of `emitClosureInstance`, returning
the expected value type instead of type erasing to Value.  This
makes it more obvious tha the value returned is an RValue (not a
BValue) which allows us to avoid a copy and use a move instead.
It would be even better to emit directly into the target, but
the code isn't wired up that way.

While here, also don't use `getIfMRValue` this way to do a
checked conversion - this silently breaks when the caller returns
another kind of value!

MODULAR_ORIG_COMMIT_REV_ID: 208c62153d161e009c8b58a4e8244d92f2f7955f