[mojo-lang] Change `__get_bvalue_as_ref` -> `__get_ref_from_value`. (#28357)
This expands the semantics of this operation to work with anything
in memory, including (non-computed) LValues, RValues, as well as
BValues, maintaining the mutability of the target in the result
type.  The name is inverted from the other __get methods we have,
but those should be changed to match when they switch away from
working on pointers.

This also unrelatedly introduces some NFC helper refactorings.

modular-orig-commit: 19d4b1a6c06762a32d088028a61c6373e87b706a