[mojo-lang] Introduce support for merging to common types. (#28359)
Add simple infra for finding a common type for coercions.  For example,
when we support classes, we'll want to support finding the
common type of two classes derived from a common base class.

More immediately, with references that share an element type, we can
merge them into a common type that is a subset of the mutability and
union of lifetimes.

Unrelatedly, this renames MojoParser/Utils.{cpp|h} to MojoUtils.*
since we have a ton of other Utils.cpp files in the tree.

MODULAR_ORIG_COMMIT_REV_ID: 32ba76cea50b619e00524f0c5edee080334439c4