[mojo-lang][CheckLifetimes] Improve precision on tracking wrt references. (#28361)
This changes CheckLifetimes to be more specific about whether an
operation is use/def/consume'ing the register value provided to it
or if the register is being dereferenced.  This different matters
in the wild world of references, because something like
`pop.select cond, ref1, ref2` or `letregdecl "x" = someref` use
the reference itself, but don't dereference it to use the underlying
value.

We still do not do proper lifetime-based tracking, but this makes
room for it.

MODULAR_ORIG_COMMIT_REV_ID: c6d82571f06c2561e0f846c4a84dfe7f287a3387