[mojo-stdlib] Vectorize overload for size as a parameter (#28374)
`algorithm.vectorize` takes `size` as an argument, so the remainder of
`size` / `simd_width` can't run in a single simd instruction as it can't
be passed as a parameter.

This has significant performance impacts for example when tiling in
matmul for smaller matrix sizes, sometimes it can make tiling
performance worse as the amount of iterations increases when there is a
remainder.

This PR adds an overload to pass size as a parameter and run the
remaining items in a single iteration, as the value is often compile
time known.

ALGORITHM_ORIG_REV_ID: 543bd8e44a513f6e7abe742d79ce3fd98cd76d9a