[KGEN] Expand `parsingStandardLibrary` docs (NFC) (#28382)
Expand the doc comment for the `ParserConfig::parsingStandardLibary`
member, and reference that in the (hidden) help text for the
`-parsing-stdlib` option.

This is an NFC because the `-parsing-stdlib` option is hidden from all
help text and documentation, so there's no observable behavioral change.

MODULAR_ORIG_COMMIT_REV_ID: 597c295897faf219a633ffa2bef999e25286f269