[mojo-parser] Slightly unhack `ASTType::getMetaType` (NFC) (#28448)
This PR slightly unwinds the hack in `ASTType::getMetaType` where the
metatype of a MetaTypeType or TraitType was itself. This combined with
`ASTType::isEqualCanon` comparing metatypes to resolve equality of two
aliased types lead to many weird issues.

In reality, this hack was to ensure that method lookup and parameter
bindings on values of a type `x.foo()` and on instances of a type like
`X.foo()` both work transparently. This PR adds a new method
`getMetaTypeOrSelf` that is only used for those cases. This still
doesn't feel like the right thing but it does improve the current
situation.

Fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: fbd01c2bb0095258e228a925a6b56badd5a7a763