[mojo] Remove unused compile options arg (NFC) (#28472)
Compilation options were added as an argument to the `PackageBuilder`
constructor in #22830, but they are no longer used as of #24840. Remove
the argument to make it clear that the the package op being built in the
constructor of this class doesn't depend on the compilation options
being passed in to `mojo`.

MODULAR_ORIG_COMMIT_REV_ID: ef90973b19cb5af7024c72f8ea997e4e42bc5b0b