[mojo] Remove obsolete comment on `ExportOp` (NFC) (#28473)
A case statement to handle `ExportOp`, and a comment explaining that
statement, existed in this code until #16457. That patch removed
`ExportOp`, but not the comment referring to it. Remove the comment to
clarify this code.

MODULAR_ORIG_COMMIT_REV_ID: bfef7297cc7736a0052cb5a951a6f1e62264d8f1