[Mojo] Replace Implementation of `isSynthetic`  (#28464)
This is a follow up to a previous PR that added an implementation of
isSynthetic for Varlets. The request is that the enum value Synthesized
be added back into VarLetDeclKind to prevent a variable from being both
synthetic and another kind.

MODULAR_ORIG_COMMIT_REV_ID: d48781ecb6e273ab4378b3e76233a73cdb91db82