[CUDA] Convey runtime-created stream into Mojo implicitly (#28486)
Part of [Removing
[Internal link]

This removes the CUDA stream from OutputChain in favor of placing it in
TLS where it can be picked up on the Mojo side by
Stream.get_current_stream(). This mirrors what we do for the Runtime
(though on the Mojo side we use Runtime() for the current or default
global runtime).

This avoids needing to plumb an OutputChainPtr from kernel entry points
down to the constrained[target == "cuda"] Mojo impls just for capturing
the stream when needed. With this I can fully remove OutputChainPtr from
Mojo except for the MOGG shims in the next PR or so.

Stream.get_current_stream() will return the null stream if called
outside of a runtime-launched kernel, eg for testing. CUDA will
interpret that to be the 'default' stream, which is correct for tests.

I reconciled myself to yet another TLS since responsibility for
constructing the stream will almost certainly be moved from the runtime
to Mojo, in which case we can remove this.

GPU_TEST_ORIG_REV_ID: 6f7ae34ac738f482fe0601a7a9e97cfefe4e5e0c