[MOGG] Use lit level metadata to replace output lambda (#28490)
Previously we just replaced anything with the default output lambda
initalizer which was not safe. Instead this uses the attached lit
metadata to target the specific lambda on the call sites.

This breaks the mouse model test which is inevitable. This requires code
which is closer to the real source than the model provides. It is still
tested by the GenericML tests however

MODULAR_ORIG_COMMIT_REV_ID: 7a89ee27717213eb0463c49fdde9c27900d3580b