[mojo][CUDA] Revert OutputChainPtr changes to fix SDK (#28556)
This reverts #28491 and #28486, since the latter specifically broke the
Mojo SDK release workflows (and the former needs to be reverted to
cleanly revert the latter).

With these changes, `mojo` fails to execute in Mojo SDK
installations on Linux; for example, from

[Internal link]

```
[mojo] Testing: `mojo build test_mandelbrot.mojo` [FAIL]
[mojo] Testing: `mojo build test_python.mojo` [FAIL]
[mojo] Testing: `mojo package` [FAIL]
[mojo] Testing: `mojo test_mandelbrot.mojo` [FAIL]
[mojo] Testing: `mojo test_python.mojo` [FAIL]
[mojo] Testing: `mojo repl` [FAIL]
[mojo] Some components may have been installed successfully, but others may not work as expected. Please submit an issue to [Internal link] include the full output of the command you just ran.
/__w/modular/modular/modular: error: failed to run script
==========================
Failure Information:
	- Script: `PATH=/__w/_tool/sccache/0.5.4/x64:/__w/modular/modular/.ninja_tracing_bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/root/.cargo/bin:/usr/local/go/bin MODULAR_HOME=/__w/modular/modular/test_modular_home PYTHONPATH=/__w/modular/modular/test_modular_home/pkg/tuff_enuff.dev.modular.com_test_6f7ae34ac738f482fe0601a7a9e97cfefe4e5e0c_packages_mojo/venv/lib/__w/modular/modular/test_modular_home/pkg/tuff_enuff.dev.modular.com_test_6f7ae34ac738f482fe0601a7a9e97cfefe4e5e0c_packages_mojo/venv/bin/python /__w/modular/modular/test_modular_home/pkg/tuff_enuff.dev.modular.com_test_6f7ae34ac738f482fe0601a7a9e97cfefe4e5e0c_packages_mojo/scripts/post-install/self-test.py --modular-home /__w/modular/modular/test_modular_home --package-path /__w/modular/modular/test_modular_home/pkg/tuff_enuff.dev.modular.com_test_6f7ae34ac738f482fe0601a7a9e97cfefe4e5e0c_packages_mojo`
```

My quick hunch is that #28486 adds some sort dependency on
CUDASupportGlobals, but that isn't being included in the Mojo SDK
installations. As a result, when the Mojo SDK is installed on a host
machine that doesn't have all of Modular's build dependencies, it fails
to execute.

While the SDK builds are broken, we lose all signal on whether we can
perform Mojo SDK (and thus MAX engine) releases, so unless there's a
known fix this needs to be reverted urgently. New commits may be landing
that further break the SDK, burying us under multiple release-blockers.

ALGORITHM_ORIG_REV_ID: 084670a0ee90ba4f5b604133fa512764f58324f8