[mojo-lang] improve unreachable code warnings for and/or expressions (#28421)
This commit checks for dead code and static branch conditions in the parser to provide better error messages than the ones provided by the LowerSemanticCF pass.  The `and` and `or` operators in particular get better messages (IE in terms of and/or instead of in terms of if).

Fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 3fe15eeb061419e60018e5062f827e35ea533245