[Support][Frameworks] Move StatsReport to a shared location (#28587)
This moves StatsReport.h/cpp to Support so that it can be shared across
frameworks.

It also removes dependency of StatsReport on LLCL (instead of depending
on LLCL runtime, it depends on TelemetryContext which is part of
Support).

Perhaps we will want to make a FrameworksSupport module in the future
and move this there.

MODULAR_ORIG_COMMIT_REV_ID: 0edf2e9ff504cf3ade418271c4ae3ca5bc759054