Revert "[Kernels] Optimize CPU reduction algorithm to reduce unnecessary ops (#28400)" (#28654)
This reverts commit a494bf489dab19b7a3c0eaf0dabce0b2995ebc89.

We determined that the previous benchmarks were getting lucky with some compiler caching. After adjusting the benchmarks to prevent the compiler from caching between benchmarks, we see worse general performance than originally published.

The issue appears to be the interaction of loop unrolling and the indices tuple passed to the input_0_fn. We suspect that adding a switch statement could solve this problem, but we lack support for this in the language.

ALGORITHM_ORIG_REV_ID: 8d31729a8379d0ffca785330d97244b9914e1e25