[mojo-lang][CheckLifetimes] Fix a harrowing subobject destruction bug. (#28638)
In a very subtle case (described in the comments) we'd fail to
upwardly propagate the full union of two if conditions when a subobject
is consumed on one side but the full object is not touched on the other
side, and would end up reporting a spurious error.

Despite the simple repro, this was encountered when tracking down
another problem, I haven't seen it reported in the wild.

MODULAR_ORIG_COMMIT_REV_ID: 66ec5c00d25379e1b69e9cb72873b4c42f38ce67