[mojo-lang][CheckLifetimes] Correct+simplify ErrorReturn handling (#28625)
Destructor insertion for `ErrorReturn` was very complicated for
historical reasons.  The invariant here is that when throwing
out of a memory-only init, 'self' doesn't need to be initialized.
We had a lot of code that was implementing the wrong thing, e.g.
in the testcase we'd run the dtor on self when all the fields
were initialized, but we should actually tear all the fields down
because we only call del when init succeeds.

This is related to issue #26048 but we still surely have other
issues with error propagation on the caller side.  This was found
when generalizing the algorithm for lifetimes, and simple is good.

MODULAR_ORIG_COMMIT_REV_ID: 7faa7542d124c9f5429038dcbe40266f0c658aa6