[SDK] Create a mojo-lldb binary (#28687)
Closes [Internal Link]

There's a current issue in which some users' PATH favors our build of
lldb instead of the system one. An easy way to fix this is to create a
copy of the lldb binary called mojo-lldb, then our config files can
point to it, and our SDK can include it instead of the original lldb
file.
I verified that a standalone SDK only has the mojo-lldb binary and not
the regular lldb.

modular-orig-commit: 0bb89dec6301257bb7cc9087a9930aa5e4094799