[Support] Add `refreshIfNecessary` to `EntitlementStore`. (#28674)
In the past, we've discussed having the EntitlementStore refresh itself based on the validity period - refreshing itself <time period> before it has expired to avoid causing the user to go through the auth flow again. Because the policy could be quite variable and application dependent, we've designed this to have the user pass a policy as a function to determine when a refresh is actually necessary.

Note that this does *not* preclude the client from calling `refresh` and forcing a refresh of the certificate if (for example) we expect that the entitlements have changed out of band.

MODULAR_ORIG_COMMIT_REV_ID: 974577a4211575bf7d584599092cee251ac0c315