[mojo-stdlib] Fix printing of empty String (#27668)
As discussed in [Internal Link] if we try
to `print(String())`, we get `(null)` as the output from `fprintf`. This is
because we're effectively calling `printf` on a `nullptr`.

Fix this by only calling `_printf` (and thus `fprintf`) if and only if
the data pointer is not null/empty by checking the length of the buffer before
calling `printf`.

modular-orig-commit: f8717d8ac798ae9db6dc79858580295d21c99d76