[LLDB] Enable python tests in CI (#28688)
It seems that the python tests haven't been running in the CI for a
while... and it seems that the issue is that the GH was setting the
`DISABLE_LLDB_PYTHON_SCRIPTING` env var as `false` even when I set it as
`FALSE`, which confused a check I have in build.py.
I suspect that the fix is just to use lower case everywhere.
I'm also updating the llvm cache key

MODULAR_ORIG_COMMIT_REV_ID: 3c99cac7f2adec6e20035f5662e8b622a4ce7b52