[mojo-lang] Error return reset consumed values (#28632)
Fixes [Internal link]

This PR is subsumed by #28625, so the only changes that will be
merged is a test case.

Here's a slightly different repro that isolates the problem a bit
better:
```
fn main() raises:
    let input: String
    with open("./in.txt", "r") as f:
        let x = f.read()
        input = "hello"
    print(input)
```

`input` is only initialized if `f.read()` doesn't raise, so if it
raises, we drop into the except case (the try-except that is de-sugared
from `with`), which error_returns out of this function.

There are therefore two paths:
1. f.read() succeeds ➡️ `input` initialized ➡️ `input` consumed
2. f.read() fails ➡️ raise inside try ➡️ error_return

Previously, error_return was not reseting the consumed state of `input`
when DestructorInsertion visits it, causing the second path to think
that `input` is consumed. This caused a discrepancy between the two
paths reaching the handle_variant for `f.read()`: one path thinks
`input` is consumed (path 2), one thinks it is not (path 1). This
resulted in handle_variant inserting a String dtor in the success path
(path 1), which is a free before init.

MODULAR_ORIG_COMMIT_REV_ID: d7e13e229d9ee695a67f5887f4e37715b58841d0