[KGEN] Fix for RaiseForLoops indvar as stride (#28623)
Fixes [Internal link]

Constant value checking is made stricter. There are now two versions:
- getValueIfConstAtLoopStart: This checks whether a value is a constant
at the start of a loop.
- getValueIfConstEveryIteration: This checks whether a value is a
constant and never changes throughout the loop.

The first one is used to find the "start" of a for loop. The second one
is used to find the "stride" & "end" of a for loop. Added tests for
cases when the stride & end are not always constant.

MODULAR_ORIG_COMMIT_REV_ID: 5b4b5763389210242bec3c4c8c268294fb26b9a6