[Support][Frameworks] Change StatsReport API and add op failures (#28791)
This makes some changes to the StatsReport API that is used to
collect op statistics during lowering for all frameworks, to
make it clearer and easier to use, and add capability to record
op failures.

Previously, countOp() would be used to record any instance of an
op encountered during lowering (whether it is lowered or not, goes
to fallback, etc). This IMO made the API confusing because
countOp and countFallback were overlapping events, so you would
always call countOp and additionally call countFallback if it
went to fallback. Besides the confusion, it also makes it harder
to use from ONNX and PT.

Instead, the API is now countLowering(), countFallback() and
countFailure(), which for a particular instance of an op only one
can occur (it is either lowered to MO, goes to fallback or fails).

MODULAR_ORIG_COMMIT_REV_ID: 4391a1595e48aaad2b5910d28cfe4b22922519da