[mojo-sdk] Don't pass in arch flags to SDK build (#28820)
As of #28631, `mojo package` no longer pre-compiles an archive to
include in a `.mojopkg` by default -- but it still does so if options
such as `-mcpu` or `-march` are specified. Remove the use of these in
the Mojo SDK build CMake when generating packages; this should make the
SDK target-agnostic, as described in #28027.

MODULAR_ORIG_COMMIT_REV_ID: b9e2873d9bc7daeb46fc108897eb20b63b640d7e