[KGEN][DebugInfo] Track SROA with DI Expression (#27882)
Continue tracking a source struct variable after the struct has been
decomposed by sroa.

This PR consists of two parts:

1. A parameterize-able version of `DIExprLeafReplacer`: Similar to the
existing `DIExprLeafReplacer`, the optimization pass can register a
leaf-replacing callback with the replacer, and use the replacer to
update DIExpressions. The difference is the plain version's callback
cannot accept any additional parameters (it can only depend on the leaf
expression type). The new version allows an opaque user-defined argument
to be passed through to the callback. This is required when additional
context is necessary for the replacing callback.

2. Using this parameterized leaf-replacer, SROA can register a
leaf-replacing callback that also accepts an index (representing the
field index in a struct). When replacing an existing leaf expression
with a struct type, it will create an `AggregatesInto` expression with
that index.

A future PR may simplify and combine the two leaf replacers into a
common version (will have to implement custom caching behavior instead
of using AttrTypeReplacer, so not doing it in this same PR).

MODULAR_ORIG_COMMIT_REV_ID: 9ce5ced6daf99af4b7b78dbb2139f33015823b42