[mojo-lang] Rework IR representation of variadics. (#29000)
Variadic arguments happen when one passes a number of values into
one dynamically-length homogenous array.  This was formerly represented
as a `!kgen.variadic<T>` value that has an operand convention matching
the original values that got passed in.

This was inconvenient for a few reasons, first it breaks the invariant
that memory types get passed by a reference.  In fact: this is a bundle
of memory types, and the bundle itself is a register value.  Further,
because the ownership was attached the the wrong thing, it was fairly
impossible to do anything with it in the projected VariadicList that
this thing gets passed to.

We solve this by always passing the `!kgen.variadic` as a borrowed
register value for variadics, and then encoding the convention of the
element type into the variadic itself, e.g.
`!kgen.variadic<T, owned_in_mem>`.  This restores the invariant that
memory arguments are always passed in registers, and exposes the
ownership to clients like VariadicList that can do something with
this.

While here, this tightens up the compiler significantly w.r.t.
variadics.  Notably, we were allowing (but late rejecting in
checklifetimes) inout and owned memory variadics, and allowing
but miscompiling owned register variadics - this is fairly terrifying
because this means we miscompiled/leaked all variadics of owned
values (e.g. PythonObjects) in defs (which default to owned).

This is a big step forward towards properly supporting both of those,
but this patch doesn't do so.

MODULAR_ORIG_COMMIT_REV_ID: 4699eb992d8ce45f3a664e9084383157d6acfec5