[mojo-lang][CheckLifetimes] Implement uninit checking for arbitrary refs. (#29004)
This verifies that the memory referenced by the uses of an indirect
reference is valid at point of use, rejects consumes of unknown
values, and adds more checking.  In addition to this, this fixes
the ownership modeling of the 'except' block in a try, and exposes
(but does not fix) that we're modeling the variant in a raising
function wrong.

MODULAR_ORIG_COMMIT_REV_ID: 500356bdbb8b48a9435386e02da0eaf88e16cc2f