[mojo-lang] Fix handling of throws + result types. NFC (#29006)
The result type of a throwing call is `variant<sometype, Error>`,
and while Error is register passable, it is not trivial - it has
a dtor that frees memory.  This changes the parser to mark all
throwing calls as having an owned result, since this error value
needs to be tracked.

In practice, this didn't affect previous code, because the parser
always generated `kgen.variadic.take` on the caller side, but this
strengthens internal invariants so we can detect other bugs.

MODULAR_ORIG_COMMIT_REV_ID: a049367e8b18d49afda12c7283b88124fc875115