[mojo-lang] Implement support for `__refitem__`/`__refattr__` (#29011)
These magic methods are an optional replacement for
`__getitem__/__setitem__` which allow intercepting the subscript or
attribute lookup syntax in Python.  These methods are implemented by
returning a reference value.

This mechanism is beneficial both for performance (because computed
LValues are more expensive in many ways than addressible LValues)
and because it allows addressing into a container in a memory safe
way - when implemented correctly the reference is bound to the
lifetime of the container.

In this patch we use it to implement trivial support for our toy
`Reference` type, enabling syntax like `aref[].noop()` which is
a tad more beautiful than `__get_value_from_ref` shenanigans.

This fixes #28427

MODULAR_ORIG_COMMIT_REV_ID: ed1d6e154b4de2aa924f49386a65d3810817cffd