[mojo-lang] Implement support for parametric mutability in references. (#29015)
One of the major problems that languages like C++ and Rust have is that
they have mutable and non-mutable references, but a lot of code needs to
apply to both mutabilities.  This leads to duplication of code, e.g.
how many times have you seen accessors duplicated like:

```
  int &getThing() { return this->field; }
  const int &getThing() const { return this->field; }
```

and of course this gets much worse at scale, xref Rusts `Index` and
`IndexMut` traits, C++'s `const_iterator` vs `iterator`, and then
everything that gets stacked on top of it.  The Mojo community has also
noticed this, e.g. with blogs like:
  https://duckki.github.io/2024/01/01/inferred-mutability.html

Mojo has a fancy-pants parameter system, so this patch switches the
mutability specifier in `!lit.ref` type to be a parameter expression
instead of a constant true/false bool.  This allows us to express
parametric mutability in addition to constant mutability, which in turn
allows users to implement algorithms and types that are generic over
mutability.

The testcases show some simple examples of this, notably extending the
prototype `Reference` type to work with both mutable and immutable
references.  The syntax used for these things is a bit ugly, but not
as critical to nail down at this point as the core model.

This fixes #28620

MODULAR_ORIG_COMMIT_REV_ID: 8fb6fef942a2395b93e3849ccd3871a9dacb810f