[mojo-lang] Rework formation of `VariadicList{Mem}` NFC (#29029)
This reworks how the compiler forms `VariadicList{Mem}` in the body
of a varargs function.  Instead of hard coding all the of the
parameter values for the generic VariadicList type, just use the
type deduction logic in the compiler.  This significantly simplifies
the implementation and moves source of truth for this stuff over to
the standard library instead of duplicating it.

This also adds a testcase for
[Internal link]
which I almost lost the code for.

MODULAR_ORIG_COMMIT_REV_ID: ca8ad52b3cd5f2c455547f11707b887fabc230c1