[mojo-lang] Expand `VariadicListMem` to work with `AnyType`. (#29033)
With the tight coupling between `VariadicListMem` and the compiler
worked out, we can now expand `VariadicListMem` to work with any
type that conforms to `AnyType`, rather than being limited to just
`AnyRegType` types.  This has the notable benefit of allowing traits
to work with varargs and gives `VariadicListMem` access to the
destructor of the type as well, which will be fun when I get back
to being able to play with references.

This fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 1165d00ac909535642a02f77b4596686e160b0eb