[mojo-lang][KGEN] Remove special case lowering for lifetimes. (#29039)
Back in June we thought that lifetimes would be pervasive parameters
on "all the things" that take a memory argument.  Because of this,
we really wanted LowerLITTypes to remove all lifetimes before
elaboration: this prevented otherwise-concrete functions from looking
parametric.

Unfortunately, the implementation of this had a significant bug: we'd
drop the parameters from the lowered generators, but we didn't update
the parameter indices in ParamIndexRefAttr's to compensate for all
the parameters after it shifting down.  As such, we'd explode late
in the compiler when a lifetime was used in a "not the last parameter"
position.

In today's world, we have an implicit lifetime feature that is used
instead.  As such, all this complexity is unneeded and can be removed.
We now just lower lifetimes to an empty kgen struct and let the
elaborator zap this all.  This is way simpler and has the benefit of
working.

This fixes #29038

MODULAR_ORIG_COMMIT_REV_ID: d28c7a338407d81f611eb7e15c8cab12a573aa1c