[mojo-stdlib] Define a `Reference` type, improve `VariadicListMem` (#29036)
This defines a starting point for a standard `Reference` type, and
uses it in the internal implementation of `VariadicListMem`. As
part of this, make subscripting a `VariadicListMem` return a
reference to the element type of the mem, enabling some nice
diffs like this:

```
 fn variadic_mem_only(*values: MemStruct) -> Int:
-    return __get_value_from_ref(values[1]).t
+    return values[1].t
```

This only applies to direct subscripts - iterating over the
`VariadicListMem` returns `Reference`'s which need to be explicitly
dereferenced with `elt[]`, but this is still a major progression.

MODULAR_ORIG_COMMIT_REV_ID: cb2546f5b8a76714899f1e2f52752e3c50944bfa