[mojo-repl] Refactor the way persistent variable types are handled (#28990)
Instead of printing the type textually, and hoping we can parse it back
in,
we now use programmatically generated aliases. This ensures that the
type is exactly correct, fixes issues for types that can't be
roundtripped via
the textual form (a lot of things with traits now), and it generally a
bit less
sketchy than the previous approach.

---------

Co-authored-by: Walter Erquinigo <walter@modular.com>
MODULAR_ORIG_COMMIT_REV_ID: 605970300ebe54fd6896456d1cb76f937443e6b7