[mojo-lang] Implement 99% support for `owned` mem varargs. (#29071)
Hot off the heels of inout varargs, this patch adds support for
`owned` memory varargs.  This allows the caller to transfer
ownership of memory objects through varargs to a callee: they then
get destroyed in the callee whenever it is done with the varargs
object.  This is progress towards #21861.

This is done by adding an "is owned" bit to `VariadicListMem` so
it knows when to destroy the objects it references in *its*
destructor.

This implementation has one extremely major flaw: the references
vended by `VariadicListMem` convey the lifetime of the source
object but do not mix in the lifetime of the `VariadicListMem`
itself, so the mem gets destroyed too early.  I will fix this in
a followup patch, but need #29069 to do it, which is pretty
orthogonal to this patch.

As part of this, this adds a `destroy_element_unsafe()` method
to `Reference`, which is a convenient analog of placement delete in
C++.

This also makes `VariadicListMem` non-copyable since it cannot
support this operation when it has owned elements. If this is important
then we either need conditional conformances or to split the type.

MODULAR_ORIG_COMMIT_REV_ID: 4a9a918199d0b82fdc814d52ccd290379eeb5acd