[Support] Use isValid in permanentPluginLibrary (#29065)
Use `isValid()` instead of checking for an error message as an
indication of whether an error has occurred in `permanentPluginLibrary`.
Previously this was erroring out when a dylib was loaded twice, since in
that case `addPermanentLibrary` returns an error message saying the
library is already loaded.
We don't want to treat the "already loaded" case as an error so check
with `.isValid()` instead.

MODULAR_ORIG_COMMIT_REV_ID: 7bf42be34e9f3fe030c59f351320df8bcfa3d1a6