[Mojo][KGEN] Fix more nits and rename in prep for keyword-only arguments (#28934)
This is another patch that renames variables and attributes to highlight
that they are positional, where ambiguity might arise in the presence of
keyword-only args/params. The patch also fixes a few other nits found
along the way, and clarifies the implementetation of some functions that
would soon change. All this is done to make subsequent non-NFC patch
reviews easier.

While doing this, it became clear that some tests in
`lit-{func,ops}.mlir` are incorrect because they use default values for
keyword-only parameters, and just happens to not crash for these
particular cases. These test are fixed, and will be improved when
keyword-only parameters are implemented. Aside from this, this patch is
NFC.

MODULAR_ORIG_COMMIT_REV_ID: 501279f320e8eaee377d309e96420ef850e85dbc