[mojo-lang] Allow 'self' in a memory type to be a `!lit.ref` (#29083)
It is typical for users to want to use normal conventions for self
like `inout` or `borrowed` or `owned`, but these conventions
intentionally hide details of how the underlying reference is passed,
notably the lifetime, but also this approach doesn't allow expression
of parametric mutability or alternate address spaces.

This patch relaxes the type checker to allow type authors to use
LOW LEVEL `!lit.ref` type for self if they choose to, which allows
advanced use cases like parametric self mutability as shown in the
testcase.  This is extraordinarily low-level though: we definitely
need a better way to do this as the reference system gets built out.

This is part of #29069, but I'm not going to call this closed until
we have something we can tell users about (this doesn't cut it).

While here, update the testcase to not need
`-mojo-experimental-lifetimes`,
with a little "parametric alias" hack. This is the only use of this
flag.

MODULAR_ORIG_COMMIT_REV_ID: 52747b5e9744a5caeddbef9dbcabd92ae2d6ad18