[KGEN] Rename AnyRegTypeType to TypeType (NFC) (#29131)
This rename is important because the KGEN-level singleton metatype
should not be opinionated about the kind of type it represents at the
LIT level. It is neither an MLIR type, or a register-passable type, or
anything else. It's just a type.

MODULAR_ORIG_COMMIT_REV_ID: 1ecfac474eea6656cf42e515d92262a1ed8a6f2d