[mojo-lang][KGEN] Add a `pop.variadic.splat` operation (#29154)
This introduces a `pop.variadic.splat` operation as requested by
the kernel team, which takes a parametric #elements to splat to.

This switches the mojo compiler to use it to make sure it is
exercised, but the real task will be for the kernel team to adopt
it where they want.  This fixes #28009.

MODULAR_ORIG_COMMIT_REV_ID: de577486819f6187c5f99990f6fe7b658992e6eb