[mojo-lang][LIT] Split + change `lit.ownership.end_lifetime` (#29170)
This splits the `lit.ownership.end_lifetime` operation into two
different ops: one for registers and one for memory, and gives the
memory one a new lifetime for the result value it is defining.

This matters because accesses through the new value should be
tracked and attributed separately from the input thing (that is,
after all the whole point of semantically ending the previous
lifetime).  This is required when we get to more lifetime-centric
tracking.

MODULAR_ORIG_COMMIT_REV_ID: db5723f177ec2c4dad47bff02b45af1ebccc8b46