[mojo-lang] Expand warnings about unneeded transfers. (#29245)
An important part of the compiler is warning about things that don't
do anything to help the user build a mental model.  In this public
issue:
[Internal Link]

A user was confused how the transfer operator worked in a case,
where... it intentionally doesn't.  We had a warning for this, but it
wasn't general enough.  We now produce:

x.mojo:4:14: warning: transfer from a value of trivial register type
'Pointer[Int, 0]' has no effect and can be removed
    takePtr(array^)
                 ^

Which makes it more clear that this is expected behavior (note,
because we don't have a good form of type sugar, we show default
parameters... sigh.)

modular-orig-commit: 5ad27d140a82bce25c85c20d3bae05dc66bbb6e5