[KGEN] Rework runtime capture of parametric closures (#29252)
This PR modifies the implementation of when a runtime closure captures a
parametric closure, and how the parametric closure's capture state is
moved onto the runtime closure's memory.

The motivation for this problem is that the current model is not
properly handling the case where the captured parametric closure itself
is parametric, because references to unbound parametric functions do not
form valid edges in the callgraph, which causes both the elaborator to
puke and does not confer the right IR to resolve-compiler-promises.

Instead, we recognize that the goal of the IR representation is not to
track the capture list of each parametric closure that could be capture,
because if they themselves are still parameterized, this cannot be known
in general, but to propagate any capture state referenced in the
`__call__` method up to the constructor, indirected through a runtime
capture state.

This PR changes the model to reflect that, and at the same time, removes
the need for `!kgen.capture_list`, since it won't be tracking a
per-parametric closure state anymore. This alone simplifies a great deal
of the implementation. We recognize that `kgen.capture_list.create` and
`kgen.capture_list.expand` are the right abstractions to work with, so
this changes resolve-compiler-promises to treat
`kgen.capture_list.expand` as a provider of all promises, and changes
`kgen.capture_list.create` to point to the `__call__` method, allowing
it to know what that total required capture state is.

This also changes the parser to generate a single `!kgen.pointer<none>`
field into the impl closure if it may capture any parametric closure,
which is a more principled approach. In the future, we can change this
to a tail allocation in the closure wrapper.

MODULAR_ORIG_COMMIT_REV_ID: 508840e7de159b46170caac356361b11abd1a5cf